pyproject.toml
[project]
name = "random-word-bot"
version = "0.1.0"
description = "A Telegram bot that sends daily vocabulary words with definitions using OpenRouter AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.52.0",
    "python-telegram-bot>=22.3",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
]
